vertexai-custom-container/requirements.txt
Flask==3.1.0
pillow==11.1.0
tensorflow==2.18.0